refactor: LoggerConfigurationKafkaExtensions class was moved to a separate class.
